+++ htdocs/error.php
@@ -61,6 +61,11 @@ require_once 'func.login.php';
     <div id="menu-share" class="menu menu-box-bottom rounded-m" data-menu-load="menu-share.html" data-menu-height="370"></div>  
     <div id="menu-colors" class="menu menu-box-bottom rounded-m" data-menu-load="menu-colors.html" data-menu-height="480"></div> 
 
+    <?php
+    que_modal_tpl('login','logout');
+    echo write_modal_tpls();
+    ?>
+
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
@@ -68,8 +73,6 @@ require_once 'func.login.php';
 
 <?php
 output_session_notification();
-que_modal_tpl('login','logout');
-echo write_modal_tpls();
 ?>
 
 </body><?php