+++ htdocs/gallery.php
@@ -304,6 +304,11 @@ foreach($items as $item){
     <div id="menu-share" class="menu menu-box-bottom rounded-m" data-menu-load="menu-share.html" data-menu-height="370"></div>  
     <div id="menu-colors" class="menu menu-box-bottom rounded-m" data-menu-load="menu-colors.html" data-menu-height="480"></div> 
 
+    <?php
+    que_modal_tpl('login','logout');
+    echo write_modal_tpls();
+    ?>
+
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
@@ -311,8 +316,6 @@ foreach($items as $item){
 
 <?php
 output_session_notification();
-que_modal_tpl('login','logout');
-echo write_modal_tpls();
 ?>
 
 </body><?php