+++ htdocs/no-reciepe.php
@@ -113,6 +113,11 @@ echo '</style>';
     <div id="menu-share" class="menu menu-box-bottom rounded-m" data-menu-load="menu-share.html" data-menu-height="370"></div>  
     <div id="menu-colors" class="menu menu-box-bottom rounded-m" data-menu-load="menu-colors.html" data-menu-height="480"></div> 
 
+    <?php
+    que_modal_tpl('login','logout');
+    echo write_modal_tpls();
+    ?>
+
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
@@ -120,8 +125,6 @@ echo '</style>';
 
 <?php
 output_session_notification();
-que_modal_tpl('login','logout');
-echo write_modal_tpls();
 ?>
 
 </body><?php