+++ htdocs/show-avatar.php
@@ -145,6 +145,11 @@ if (($items = $mysqli->prepared_query("SELECT * FROM `" . $kista_dp . "replicate
     <div id="menu-share" class="menu menu-box-bottom rounded-m" data-menu-load="menu-share.html" data-menu-height="370"></div>  
     <div id="menu-colors" class="menu menu-box-bottom rounded-m" data-menu-load="menu-colors.html" data-menu-height="480"></div> 
 
+    <?php
+    que_modal_tpl('login','logout');
+    echo write_modal_tpls();
+    ?>
+
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
@@ -152,8 +157,6 @@ if (($items = $mysqli->prepared_query("SELECT * FROM `" . $kista_dp . "replicate
 
 <?php
 output_session_notification();
-que_modal_tpl('login','logout');
-echo write_modal_tpls();
 ?>
 
 </body><?php