+++ htdocs/upload-face.php
@@ -324,16 +324,15 @@ switch($modulus) {
     <div id="menu-share" class="menu menu-box-bottom rounded-m" data-menu-load="menu-share.html" data-menu-height="370"></div>  
     <div id="menu-colors" class="menu menu-box-bottom rounded-m" data-menu-load="menu-colors.html" data-menu-height="480"></div> 
 
+    <?php
+    que_modal_tpl('login','logout');
+    echo write_modal_tpls();
+    ?>
+
 </div>
 
 <script type="text/javascript" src="scripts/bootstrap.min.js"></script>
 <script type="text/javascript" src="scripts/custom.js.php?<?=$html_NoCache_Version?>"></script>
-
-<?php
-output_session_notification();
-que_modal_tpl('login','logout');
-echo write_modal_tpls();
-?>
 
 <script>
 
@@ -415,10 +414,11 @@ document.addEventListener("DOMContentLoaded", function() {
   });
 });
 
-
-
 </script>
 
+<?php
+output_session_notification();
+?>
 
 </body><?php
 ob_end_flush();