+++ htdocs/docs.php
@@ -1,0 +1,92 @@
+<?php
+
+
+ob_start();
+session_cache_expire(720);
+session_start();
+
+define('APPDATA_PATH', dirname(__FILE__) . '/inc_appdata');
+define('UPLOAD_PATH', dirname(__FILE__) . '/uploaded_files');
+
+require_once 'func.inc.php';
+require_once 'func.login.php';
+
+?>
+<!DOCTYPE HTML>
+<html lang="<?=$lang?>">
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+<meta name="apple-mobile-web-app-capable" content="yes">
+<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
+<meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1, viewport-fit=cover" />
+<title><?=$appConf['headTitle']?></title>
+<link rel="stylesheet" type="text/css" href="styles/bootstrap.css">
+<link rel="stylesheet" type="text/css" href="styles/style.css?<?=$html_NoCache_Version?>">
+<link href="https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap" rel="stylesheet">
+<link rel="stylesheet" type="text/css" href="fonts/css/fontawesome-all.min.css">
+<link rel="manifest" href="_manifest.json.php?<?=$PWA_APP_VER?>" data-pwa-version="<?=$PWA_APP_VER?>">
+<link rel="apple-touch-icon" sizes="180x180" href="<?=$PWA_LANG['ico_folder']?>/icon-192x192.png">
+<?=$PWA_LANG['highlight']?>
+</head>
+
+<body class="theme-light">
+
+<div id="preloader" class="preloader-hide"><div class="spinner-border color-highlight" role="status"></div></div>
+
+<div id="page">
+
+    <div class="header header-fixed header-logo-center">
+        <a href="error.php" class="header-title">Doc</a>
+        <?=$HTML_HEADER('header-fixed')?>
+    </div>
+
+    <?=$HTML_FOOTER(3)?>
+
+
+    <div class="page-content header-clear-medium">
+
+            <div class="card card-style">
+                <div class="content">
+                    <p class="mb-n1 color-highlight font-600">Doc</p>
+<?php
+
+$doc = '' . _GET('doc') . '.md';
+if( file_exists('./docs/' . $doc) ){
+    $text = file_get_contents('./docs/' . $doc);
+    $Parsedown = new Parsedown();
+    echo $Parsedown->text( $text );
+} else {
+    $text = '# not found';
+    $Parsedown = new Parsedown();
+    echo $Parsedown->text( $text );
+}
+
+?>
+               </div>
+            </div>
+
+        <div data-menu-load="<?=$appConf['menuFooter' . $_menuSuffix]?>"></div>
+    </div>
+    <!-- Page content ends here-->
+    
+    <div id="menu-main" class="menu menu-box-left rounded-0" data-menu-load="menu-main.html" data-menu-width="280" data-menu-active="nav-media"></div>
+    <div id="menu-share" class="menu menu-box-bottom rounded-m" data-menu-load="menu-share.html" data-menu-height="370"></div>  
+    <div id="menu-colors" class="menu menu-box-bottom rounded-m" data-menu-load="menu-colors.html" data-menu-height="480"></div> 
+
+    <?php
+    que_modal_tpl('login','logout');
+    echo write_modal_tpls();
+    ?>
+
+</div>
+
+<script type="text/javascript" src="scripts/bootstrap.min.js"></script>
+<script type="text/javascript" src="scripts/custom.js.php?<?=$html_NoCache_Version?>"></script>
+
+<?php
+output_session_notification();
+?>
+
+</body><?php
+ob_end_flush();
+?>