+++ htdocs/func.inc.php
@@ -194,9 +194,9 @@ function HTML_FOOTER_AVATAR($i){
     global $lang;
 
     if($lang == 'nb')
-        $titles = ['Sider', 'Last opp', 'Galleri'];
+        $titles = ['Hjem', 'Last opp', 'Galleri'];
         else
-        $titles = ['Pages', 'Upload', 'Gallery'];
+        $titles = ['Home', 'Upload', 'Gallery'];
 
     return '
     <div id="footer-bar" class="footer-bar-6">
@@ -204,7 +204,7 @@ function HTML_FOOTER_AVATAR($i){
         <a href="index.php" class="' . setActiveIf($i,2) . '"><i class="fa fa-file"></i><span>' . $titles[0] . '</span></a>
         <a href="upload-face.php" class="circle-nav ' . setActiveIf($i,3) . ' external-link"><i class="fa fa-home"></i><span>' . $titles[1] . '</span></a>
         <a href="gallery-avatar.php" class="' . setActiveIf($i,4) . '"><i class="fa fa-camera"></i><span>' . $titles[2] . '</span></a>
-        <a href="#" data-menu="menu-main" class="' . setActiveIf($i,5) . '"><i class="fa fa-bars"></i><span>Menu</span></a>
+        <a href="#" data-menu="menu-main" data-menu-active="nav-components" class="' . setActiveIf($i,5) . '"><i class="fa fa-bars"></i><span>Menu</span></a>
     </div>
     ';
 }