+++ htdocs/gallery-avatar.php
@@ -65,6 +65,8 @@ if($lang == 'en'){
 
 
 
+
+        <!--
         <div class="card card-style">
             <div class="content mt-3">
                 <p class="mb-n1 color-highlight font-600 mb-n1">Project Briefing </p>
@@ -73,10 +75,11 @@ if($lang == 'en'){
                     echo 'USER_ID: ' . $_SESSION['USER_ID'] . '<br>';
                     $user_google_id = $_SESSION['USER_GOOGLE_LOGIN'][0] ?? 0;
                     echo 'user_google_id: ' . $user_google_id . '<br>';
+                    echo htmlentities( "SELECT *, `reim`.`filename` AS `filename` FROM `" . $kista_dp . "replicate__images` `reim` INNER JOIN `" . $kista_dp . "replicate__uploads` `reup` ON `reim`.reid = `reup`.reid AND `reup`.user_id = ?" );
                 ?></p>
             </div>
 		</div>
-
+        -->
 
         <div class="content mt-0 mb-0">
             <div class="d-flex">
@@ -90,10 +93,30 @@ if($lang == 'en'){
         </div>
         
 <?php
-
-
 define('REPLICATE_INFERENCE_FOLDER', UPLOAD_PATH . '/ri');
-if (($items = $mysqli->prepared_query("SELECT *, `reim`.`filename` AS `filename` FROM `" . $kista_dp . "replicate__images` `reim` INNER JOIN `" . $kista_dp . "replicate__uploads` `reup` ON `reim`.reid = `reup`.reid AND `reup`.user_id = ?", 'i', [$USER_ID])) !== []) {
+
+$user_google_id = $_SESSION['USER_GOOGLE_LOGIN'][0] ?? 0;
+if($user_google_id){
+    $p_sql = [
+        "SELECT *, `reim`.`filename` AS `filename` 
+         FROM `kistaai_replicate__images` `reim` 
+         INNER JOIN `kistaai_replicate__uploads` `reup` ON `reim`.`reid` = `reup`.`reid` 
+         INNER JOIN `kistaai_users__sessions` `s` ON `reup`.`user_id` = `s`.`user_id` 
+         WHERE (`s`.`user_id` = ? OR `s`.`google_id` = ?)", 
+        'ii', 
+        [$USER_ID, $user_google_id]
+    ];
+} else {
+    $p_sql = [
+        "SELECT *, `reim`.`filename` AS `filename` 
+         FROM `" . $kista_dp . "replicate__images` `reim` 
+         INNER JOIN `" . $kista_dp . "replicate__uploads` `reup` ON `reim`.reid = `reup`.reid AND `reup`.user_id = ?", 
+        'i', 
+        [$USER_ID]
+    ];
+}
+
+if (($items = $mysqli->prepared_query($p_sql)) !== []) {
 
     echo '
         <div data-splide=\'{"autoplay":false}\' class="splide double-slider visible-slider slider-no-arrows slider-no-dots" id="double-slider-1">