+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Added, Changed, Deprecated, Removed, Fixed, Securit
 - changed, moving off-canvas elements outside page-content
 - added, help section explaining avatarify and ai-jargon
 - changed, new Avatarify main logo
+- changed, avatarify gallery queries now handles google login sessions
+- fixed, when logging in with google the uses session table was not updated
 
 ### v0.11.4
 - fixed, index had old links to kista-ai.steinhaug.no, now set in _vars.php